pyproject.toml
[project]
name = "pythonlessons2024-ai"
version = "0.1.0"
description = ""
authors = [
    {name = "Вадим",email = "pvs.versia@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "g4f (>=0.5.3.2,<0.6.0.0)",
    "pydantic-ai>=0.2.14"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
